Code review fixes - ActionManager now split into base class and loader specific version
